add bulk delete warning for delete methods on wp core models

+++ src/Models/WPComment.php
@@ -223,6 +223,10 @@ class WPComment extends Model
             $ids = $this->getID();
         }
 
+        if(is_array($ids)) {
+            throw new ModelException(static::class . ' not deleted: bulk deleting not supported due to WordPress performance issues.');
+        }
+
         do_action('typerocket_model_delete', $this, $ids);
 
         $delete = wp_delete_comment($ids);
