+++ src/Models/WPTerm.php
@@ -357,6 +357,10 @@ class WPTerm extends Model
             $ids = $this->getID();
         }
 
+        if(is_array($ids)) {
+            throw new ModelException(static::class . ' not deleted: bulk deleting not supported due to WordPress performance issues.');
+        }
+
         do_action('typerocket_model_delete', $this, $ids);
 
         $delete = wp_delete_term($ids, $this->wpTerm($ids)->taxonomy);
